change the type of length_of_audio

+++ db/migrate/20141015090807_add_length_of_audio_to_episodes.rb
@@ -1,5 +1,5 @@
 class AddLengthOfAudioToEpisodes < ActiveRecord::Migration
   def change
-    add_column :episodes, :length_of_audio, :float, default: 0
+    add_column :episodes, :length_of_audio, :integer, default: 0
   end
 end
